# docs/api/api-docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FenixTech API",
    "description": "API para la gestión de la plataforma FenixTech, incluyendo categorías, subcategorías, productos, pedidos y usuarios.",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "name": "FenixTech Support",
      "email": "support@fenixtech.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://springdoc.org"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost/fenixtech/api/v1",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/subcategories/{id}": {
      "get": {
        "tags": [
          "Subcategories"
        ],
        "summary": "Obtener subcategoría por ID",
        "description": "Devuelve una subcategoría por su ID",
        "operationId": "findById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Subcategoría obtenida con éxito",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Subcategories"
                }
              }
            }
          },
          "404": {
            "description": "Subcategoría no encontrada",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Subcategories"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Subcategories"
        ],
        "summary": "Actualizar subcategoría por ID",
        "description": "Actualiza una subcategoría por su ID",
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubcategoriesRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Subcategoría actualizada con éxito",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Subcategories"
                }
              }
            }
          },
          "404": {
            "description": "Subcategoría no encontrada",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Subcategories"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Subcategories"
        ],
        "summary": "Eliminar subcategoría por ID",
        "description": "Elimina una subcategoría por su ID",
        "operationId": "deleteById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Subcategoría eliminada con éxito"
          },
          "404": {
            "description": "Subcategoría no encontrada"
          }
        }
      }
    },
...
    "/shipments/{id}": {
      "get": {
        "tags": [
          "Shipments"
        ],
        "summary": "Obtener envío por ID",
        "description": "Devuelve un envío por su ID",
        "operationId": "findById_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Envío obtenido con éxito",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ShipmentResponseDTO"
                }
              }
            }
          },
          "404": {
            "description": "Envío no encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ShipmentResponseDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Shipments"
        ],
        "summary": "Actualizar envío por ID",
        "description": "Actualiza un envío por su ID",
        "operationId": "update_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShipmentUpdateCarrierDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Envío actualizado con éxito",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Shipments"
                }
              }
            }
          },
          "404": {
            "description": "Envío no encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Shipments"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Shipments"
        ],
        "summary": "Eliminar envío por ID",
        "description": "Elimina un envío por su ID",
        "operationId": "deleteById_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Envío eliminado con éxito"
          },
          "404": {
            "description": "Envío no encontrado"
          }
        }
      }
    },
    "/reviews/{id}": {
      "put": {
        "tags": [
          "Reviews"
        ],
        "summary": "Actualizar una review",
        "description": "Actualiza una review por su ID",
        "operationId": "update_3",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewsRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Review actualizada con éxito",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Reviews"
                }
              }
            }
          },
          "404": {
            "description": "Review no encontrada",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Reviews"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Reviews"
        ],
        "summary": "Eliminar una review",
        "description": "Elimina una review por su ID",
        "operationId": "deleteById_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Review eliminada con éxito"
          },
          "404": {
            "description": "Review no encontrada"
          }
        }
      }
    },
    "/proposals/{id}": {
      "get": {
        "tags": [
          "Proposals"
        ],
        "summary": "Obtener propuesta por ID",
        "description": "Devuelve una propuesta por su ID",
        "operationId": "findProposalById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Propuesta obtenida con éxito",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Proposals"
                }
              }
            }
          },
          "404": {
            "description": "Propuesta no encontrada",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Proposals"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Proposals"
        ],
        "summary": "Actualizar propuesta (Uso ADMIN)",
        "description": "Permite al administrador modificar el estado, título, descripción o categoría",
        "operationId": "update_4",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalRequestUpdateDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Propuesta actualizada con éxito",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Proposals"
                }
              }
            }
          },
          "404": {
            "description": "Propuesta no encontrada",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Proposals"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Proposals"
        ],
        "summary": "Eliminar una propuesta",
        "description": "Elimina una propuesta por su ID",
        "operationId": "deleteById_3",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Propuesta eliminada con éxito"
          },
          "404": {
            "description": "Propuesta no encontrada"
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Obtener producto por ID",
        "description": "Devuelve un producto por su ID",
        "operationId": "findById_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Producto obtenido con éxito",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Products"
                }
              }
            }
          },
          "404": {
            "description": "Producto no encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Products"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Products"
        ],
        "summary": "Actualizar un producto",
        "description": "Actualiza un producto existente",
        "operationId": "update_5",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductRequestUpdateDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Producto actualizado con éxito",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Products"
                }
              }
            }
          },
          "404": {
            "description": "Producto no encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Products"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Products"
        ],
        "summary": "Eliminar producto por ID",
        "description": "Elimina un producto por su ID",
        "operationId": "deleteById_4",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Producto eliminado con éxito"
          },
          "404": {
            "description": "Producto no encontrado"
          }
        }
      }
    },
    "/posts/{id}": {
      "get": {
        "tags": [
          "Posts"
        ],
        "summary": "Obtener post por ID",
        "description": "Devuelve un post por su ID",
        "operationId": "findPostById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Post obtenido con éxito",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Posts"
                }
              }
            }
          },
          "404": {
            "description": "Post no encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Posts"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Posts"
        ],
        "summary": "Actualizar un post",
        "description": "Actualiza un post existente",
        "operationId": "updatePost",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostsRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Post actualizado con éxito",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Posts"
                }
              }
            }
          },
          "404": {
            "description": "Post no encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Posts"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Posts"
        ],
        "summary": "Eliminar post por ID",
        "description": "Elimina un post por su ID",
        "operationId": "deleteById_5",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Post eliminado con éxito"
          },
          "404": {
            "description": "Post no encontrado"
          }
        }
      }
    },
    "/orders/{id}/status": {
      "put": {
        "tags": [
          "Orders"
        ],
        "summary": "Actualizar estado del pedido",
        "description": "Cambia el estado logístico y gestiona el stock en caso de cancelación.",
        "operationId": "updateStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderRequestUpdateDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Estado actualizado con éxito",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Orders"
                }
              }
            }
          },
          "400": {
            "description": "Cambio de estado no permitido",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Orders"
                }
              }
            }
          },
          "404": {
            "description": "Pedido no encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Orders"
                }
              }
            }
          }
        }
      }
    },
    "/companies/{id}": {
      "get": {
        "tags": [
          "Companies"
        ],
        "summary": "Obtener empresa activa por ID",
        "description": "Busca una empresa por su ID siempre que esté activa",
        "operationId": "findByIdActive",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Empresa encontrada con éxito",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Companies"
                }
              }
            }
          },
          "404": {
            "description": "Empresa no encontrada o inactiva",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Companies"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Companies"
        ],
        "summary": "Actualizar empresa por ID",
        "description": "Actualiza los datos básicos de la empresa",
        "operationId": "update_6",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompaniesRequestUpdateDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Empresa actualizada con éxito",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Companies"
                }
              }
            }
          },
          "404": {
            "description": "Empresa no encontrada",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Companies"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Companies"
        ],
        "summary": "Borrado lógico de empresa",
        "description": "Desactiva la empresa y su usuario asociado",
        "operationId": "deleteById_6",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Empresa desactivada correctamente"
          },
          "404": {
            "description": "Empresa no encontrada"
          }
        }
      }
    },
...
    "/categories/{id}": {
      "get": {
        "tags": [
          "Categories"
        ],
        "summary": "Obtener categoría por ID",
        "description": "Devuelve una categoría por su ID",
        "operationId": "findById_3",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Categoría obtenida con éxito",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Categories"
                }
              }
            }
          },
          "404": {
            "description": "Categoría no encontrada",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Categories"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Categories"
        ],
        "summary": "Actualizar una categoría por ID",
        "description": "Actualiza una categoría por su ID",
        "operationId": "update_8",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoriesRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Categoría actualizada con éxito",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Categories"
                }
              }
            }
          },
          "404": {
            "description": "Categoría no encontrada",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Categories"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Categories"
        ],
        "summary": "Eliminar una categoría por ID",
        "description": "Elimina una categoría por su ID",
        "operationId": "deleteById_7",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Categoría eliminada con éxito"
          },
          "404": {
            "description": "Categoría no encontrada"
          }
        }
      }
    },
...
    "/shipments": {
      "get": {
        "tags": [
          "Shipments"
        ],
        "summary": "Obtener todos los envíos",
        "description": "Devuelve una lista de todos los envíos",
        "operationId": "findAllShipments",
        "responses": {
          "200": {
            "description": "Envíos obtenidos con éxito",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Shipments"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Shipments"
        ],
        "summary": "Añadir envío",
        "description": "Añade un nuevo envío",
        "operationId": "save_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShipmentRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Envío creado con éxito",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Shipments"
                }
              }
            }
          },
          "404": {
            "description": "Pedido no encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Shipments"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Subcategories": {
        "type": "object",
        "description": "Modelo de Subcategorías",
        "properties": {
          "subcategoryId": {
            "type": "integer",
            "format": "int32",
            "description": "Identificador único de la subcategoría",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Nombre de la subcategoría",
            "example": "Laptops",
            "minLength": 1
          },
          "description": {
            "type": "string",
            "description": "Descripción de la subcategoría",
            "example": "Ordenadores portátiles de diversas marcas"
          },
          "isActive": {
            "type": "boolean",
            "description": "Visibilidad de la subcategoría",
            "example": true
          },
          "category": {
            "$ref": "#/components/schemas/Categories"
          },
          "products": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Products"
            }
          }
        },
        "required": [
          "name"
        ]
      },
...
      "ShipmentUpdateCarrierDTO": {
        "type": "object",
        "properties": {
          "carrierId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "required": [
          "carrierId"
        ]
      },
...
    }
  }
}